# docs/stackr-api/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Stackr Catalogue API
  version: 1.0.0
  summary: Stable public API for catalogue, delta sync and search.
servers:
  - url: https://api.stackr.app/v1
    description: Production edge gateway
  - url: http://localhost:8787/v1
    description: Local Cloudflare Worker
...
paths:
...
  /catalog/manifest:
    get:
      tags: [Catalogue]
      operationId: getCatalogManifest
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Current catalogue sync manifest.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ManifestResponse"
        "304":
          description: Manifest has not changed.
  /catalog/delta:
    get:
      tags: [Catalogue]
      operationId: getCatalogDelta
      parameters:
        - name: since
          in: query
          schema:
            type: integer
            minimum: 0
          description: Last applied catalogue change sequence.
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Cursor-paginated catalogue inserts, updates and deprecations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeltaResponse"
        "304":
          description: Delta page has not changed.
        "400":
          $ref: "#/components/responses/Error"
  /languages:
    get:
      tags: [Catalogue]
      operationId: listLanguages
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
      responses:
        "200":
          description: Supported catalogue languages.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LanguagesResponse"
  /series:
    get:
      tags: [Catalogue]
      operationId: listSeries
      parameters:
        - $ref: "#/components/parameters/Language"
        - name: game
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Cursor-paginated catalogue series.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeriesResponse"
  /sets:
    get:
      tags: [Catalogue]
      operationId: listSets
      parameters:
        - $ref: "#/components/parameters/Language"
        - name: game
          in: query
          schema:
            type: string
        - name: seriesId
          in: query
          schema:
            type: string
            format: uuid
        - name: setCode
          in: query
          schema:
            type: string
        - name: region
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Cursor-paginated catalogue sets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetsResponse"
  /sets/{setId}:
    get:
      tags: [Catalogue]
      operationId: getSet
      parameters:
        - $ref: "#/components/parameters/SetId"
      responses:
        "200":
          description: A canonical set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetResponse"
        "404":
          $ref: "#/components/responses/Error"
  /sets/{setId}/cards:
    get:
      tags: [Catalogue]
      operationId: listSetCards
      parameters:
        - $ref: "#/components/parameters/SetId"
        - $ref: "#/components/parameters/Language"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Cursor-paginated cards in a canonical set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardsResponse"
  /cards/{cardId}:
    get:
      tags: [Catalogue]
      operationId: getCard
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: A canonical card printing with variants.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "404":
          $ref: "#/components/responses/Error"
  /cards/{cardId}/variants:
    get:
      tags: [Catalogue]
      operationId: listCardVariants
      parameters:
        - $ref: "#/components/parameters/CardId"
      responses:
        "200":
          description: Variants for a canonical card printing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardVariantsResponse"
  /cards/{variantId}/price:
    get:
      tags: [Market]
      operationId: getCardVariantPrice
      parameters:
        - $ref: "#/components/parameters/VariantId"
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/Condition"
        - $ref: "#/components/parameters/Grader"
        - $ref: "#/components/parameters/Grade"
      responses:
        "200":
          description: Latest provider-neutral price estimate for one exact variant/product identity.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardPriceResponse"
        "400":
          $ref: "#/components/responses/Error"
  /cards/{variantId}/price-history:
    get:
      tags: [Market]
      operationId: getCardVariantPriceHistory
      parameters:
        - $ref: "#/components/parameters/VariantId"
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/Condition"
        - $ref: "#/components/parameters/Grader"
        - $ref: "#/components/parameters/Grade"
        - name: observationType
          in: query
          schema:
            type: string
            enum: [sold_observation, active_listing]
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Public-safe pricing evidence history with active and sold rows explicitly labelled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceHistoryResponse"
        "400":
          $ref: "#/components/responses/Error"
  /market/movers:
    get:
      tags: [Market]
      operationId: listMarketMovers
      parameters:
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Market identities with recent estimate movement.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketMoversResponse"
  /market/opportunities:
    get:
      tags: [Market]
      operationId: listMarketOpportunities
      parameters:
        - $ref: "#/components/parameters/ProductType"
        - $ref: "#/components/parameters/Currency"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Active listings below exact-variant estimates. These are not sold comps.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketOpportunitiesResponse"
  /search:
    get:
      tags: [Search]
      operationId: searchCatalog
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 160
        - $ref: "#/components/parameters/Language"
        - name: setId
          in: query
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Search results with explicit match reasons.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          $ref: "#/components/responses/Error"
  /assets/manifest:
    get:
      tags: [Assets]
      operationId: getAssetManifest
      parameters:
        - name: assetType
          in: query
          schema: { type: string }
        - name: setId
          in: query
          schema: { type: string, format: uuid }
        - name: printingId
          in: query
          schema: { type: string, format: uuid }
        - name: variantId
          in: query
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Approved public catalogue asset manifest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Envelope"
...
components:
  securitySchemes:
    SupabaseBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived Supabase user access token. Service-role keys are forbidden.
  headers:
...
    ETag:
      schema:
        type: string
    CacheControl:
      schema:
        type: string
...
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema:
        type: string
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
      description: Opaque cursor returned by the previous page.
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
    Language:
      name: language
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/LanguageCode"
    SetId:
      name: setId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CardId:
      name: cardId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VariantId:
      name: variantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProductType:
      name: productType
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/MarketProductType"
    Currency:
      name: currency
      in: query
      required: false
      schema:
        type: string
        pattern: "^[A-Z]{3}$"
        default: GBP
    Condition:
      name: condition
      in: query
      required: false
      schema:
        type: string
    Grader:
      name: grader
      in: query
      required: false
      schema:
        type: string
    Grade:
      name: grade
      in: query
      required: false
      schema:
        type: string
  responses:
    Error:
      description: Structured API error envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    LanguageCode:
      type: string
      enum: [en, ja, zh-Hans, zh-Hant, ko]
    MarketProductType:
      type: string
      enum: [raw_card, graded_card, sealed_product]
    MarketEvidenceStatus:
      type: string
      enum: [recent_sold_value, thin_sold_value, market_estimate, asking_price_indication, unavailable]
    EnvelopeMeta:
      type: object
      required: [requestId, apiVersion, generatedAt]
      properties:
        requestId:
          type: string
        apiVersion:
          type: string
          const: "1"
        generatedAt:
          type: string
          format: date-time
        pagination:
          $ref: "#/components/schemas/Pagination"
    Pagination:
      type: object
      required: [limit, nextCursor]
      properties:
        limit:
          type: integer
        nextCursor:
          type: [string, "null"]
...
    ManifestResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required:
                - currentCatalogueVersion
                - minCompatibleAppSchemaVersion
                - latestChangeSequence
                - availableLanguageShards
                - assetBaseUrl
                - modelIndexVersion
                - generatedAt
                - etag
              properties:
                currentCatalogueVersion:
                  type: string
                catalogueVersionId:
                  type: [string, "null"]
                  format: uuid
                minCompatibleAppSchemaVersion:
                  type: string
                latestChangeSequence:
                  type: integer
                availableLanguageShards:
                  type: array
                  items:
                    type: object
                assetBaseUrl:
                  type: [string, "null"]
                modelIndexVersion:
                  type: [string, "null"]
                generatedAt:
                  type: string
                  format: date-time
                etag:
                  type: string
    DeltaResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    LanguagesResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    SeriesResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    SetsResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    SetResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    CardsResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    CardResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    CardVariantsResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    CardPriceResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [variantId, productType, currency, status, estimates, sample, confidence]
              properties:
                variantId:
                  type: string
                  format: uuid
                productType:
                  $ref: "#/components/schemas/MarketProductType"
                currency:
                  type: string
                status:
                  $ref: "#/components/schemas/MarketEvidenceStatus"
                priceType:
                  $ref: "#/components/schemas/MarketEvidenceStatus"
                estimates:
                  type: object
                  properties:
                    low:
                      type: [number, "null"]
                    central:
                      type: [number, "null"]
                    high:
                      type: [number, "null"]
                sample:
                  type: object
                confidence:
                  type: object
                fallbackEstimate:
                  type: [object, "null"]
                unavailableReason:
                  type: [string, "null"]
    PriceHistoryResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    MarketMoversResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    MarketOpportunitiesResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
    SearchResponse:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              required: [query, normalizedQuery, results]
              properties:
                query:
                  type: string
                normalizedQuery:
                  type: string
                results:
                  type: array
                  items:
                    $ref: "#/components/schemas/SearchResult"
    SearchResult:
      type: object
      required: [type, reason]
      properties:
        type:
          type: string
          enum: [card, set]
        reason:
          type: string
          enum:
            - exact_canonical_id
            - exact_external_id
            - exact_set_code_collector_number
            - exact_collector_number
            - exact_collector_number_in_set
            - exact_name_in_set
            - exact_name
            - exact_alias
            - exact_translated_name
            - fuzzy_name
    Envelope:
      type: object
      required: [data, meta]
      properties:
        data:
          type: object
        meta:
          $ref: "#/components/schemas/EnvelopeMeta"
    ErrorEnvelope:
      type: object
      required: [error, meta]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details: {}
        meta:
          type: object